+++ secant/Assets.xcassets/Logos/Contents.json
